Fixed index naming mismatch introduced in the refactor
